+++ walkability/components/tactile_paving/tactile_paving_artifact.py
@@ -2,7 +2,14 @@ import logging
 from pathlib import Path
 
 import geopandas as gpd
-from climatoology.base.artifact_creators import Artifact, ArtifactMetadata, Legend, create_vector_artifact
+import plotly.graph_objects as go
+from climatoology.base.artifact_creators import (
+    Artifact,
+    ArtifactMetadata,
+    Legend,
+    create_plotly_chart_artifact,
+    create_vector_artifact,
+)
 from climatoology.base.computation import ComputationResources
 
 from walkability.components.utils.misc import Topics, generate_colors, get_tactile_paving_legend
@@ -20,7 +27,7 @@ def build_tactile_paving_artifact(tactile_locations: gpd.GeoDataFrame, resources
         data=tactile_locations[['osm_id', 'osm_type', 'color', 'label', 'geometry']],
         metadata=ArtifactMetadata(
             name='Tactile Paving',
-            summary='Does this path have tactile paving?',
+            summary='Presence of tactile paving at pedestrian crossings (displayed as points), platforms, and stairs',
             description=Path('resources/components/tactile_paving/tactile_paving_description.md').read_text(),
             filename='tactile_paving',
             primary=False,
@@ -28,4 +35,18 @@ def build_tactile_paving_artifact(tactile_locations: gpd.GeoDataFrame, resources
         ),
         resources=resources,
         legend=Legend(legend_data=get_tactile_paving_legend()),
+    )
+
+
+def build_tactile_paving_chart_artifact(figure: go.Figure, resources: ComputationResources) -> Artifact:
+    return create_plotly_chart_artifact(
+        figure=figure,
+        metadata=ArtifactMetadata(
+            name='Distribution of Tactile Paving Categories',
+            summary='What percentage of crossings, platforms, and stairs have tactile paving?',
+            filename='tactile_paving_summary',
+            primary=False,
+            tags={Topics.SAFETY},
+        ),
+        resources=resources,
     )